test_st_when_fetcher_crashes is moved from skvbc_persistence to skvbc_state_transfer suit
