+++ tests/apollo/test_skvbc_persistence.py
@@ -168,82 +168,6 @@ class SkvbcPersistenceTest(unittest.TestCase):
 
         await skvbc.read_your_writes()
 
-    @with_trio
-    @with_bft_network(start_replica_cmd)
-    @verify_linearizability()
-    async def test_st_when_fetcher_crashes(self, bft_network, tracker):
-        """
-        Start N-1 nodes out of a N node cluster (hence 1 stale node). Write a specific key,
-        then enough data to the cluster to trigger a checkpoint. Then stop all the nodes,
-        restart the N-1 nodes that should have checkpoints as well as the stale node.
-
-        Kill and restart the stale node multiple times during fetching and
-        make sure that it catches up.
-
-        Then force a quorum including the stale node, and try to read the specific
-        key.
-
-        After that ensure that a newly put value can be retrieved.
-        """
-        stale_node = random.choice(bft_network.all_replicas(without={0}))
-        skvbc = kvbc.SimpleKVBCProtocol(bft_network, tracker)
-
-        client, known_key, known_val = \
-            await skvbc.prime_for_state_transfer(stale_nodes={stale_node})
-
-        # Start the empty replica, wait for it to start fetching, then stop
-        # it.
-        bft_network.start_replica(stale_node)
-        await bft_network.wait_for_fetching_state(stale_node)
-        bft_network.stop_replica(stale_node)
-
-        # Loop repeatedly starting and killing the destination replica after
-        # state transfer has started. On each restart, ensure the node is
-        # still fetching or that it has received all the data.
-        await self._fetch_or_finish_state_transfer_while_crashing(bft_network, 0, stale_node)
-
-        # Restart the replica and wait for state transfer to stop
-        bft_network.start_replica(stale_node)
-        await bft_network.wait_for_state_transfer_to_stop(0, stale_node)
-
-        await bft_network.force_quorum_including_replica(stale_node)
-
-        # Retrieve the value we put first to ensure state transfer worked
-        # when the log went away
-        kvpairs = await skvbc.send_read_kv_set(client, known_key)
-        self.assertDictEqual(dict([(known_key, known_val)]), kvpairs)
-
-        # Perform a put/get transaction pair to ensure we can read newly
-        # written data after state transfer.
-
-        await skvbc.read_your_writes()
-
-    async def _fetch_or_finish_state_transfer_while_crashing(self,
-                                                             bft_network,
-                                                             up_to_date_node,
-                                                             stale_node,
-                                                             nb_crashes=20):
-       for _ in range(nb_crashes):
-           log.log_message(message_type=f'Restarting replica {stale_node}')
-           bft_network.start_replica(stale_node)
-           try:
-               await bft_network.wait_for_fetching_state(stale_node)
-               # Sleep a bit to give some time for the fetch to make progress
-               await trio.sleep(random.uniform(0, 1))
-
-           except trio.TooSlowError:
-               # We never made it to fetching state. Are we done?
-               try:
-                    await bft_network.wait_for_state_transfer_to_stop(
-                       up_to_date_node, stale_node)
-                    break
-               except trio.TooSlowError:
-                   self.fail("State transfer did not complete, " +
-                             "but we are not fetching either!")
-           finally:
-               log.log_message(message_type=f'Stopping replica {stale_node}')
-               bft_network.stop_replica(stale_node)
-
     @skip_for_tls
     @unittest.skip("Fails because of BC-7264")
     @with_trio
@@ -487,7 +411,7 @@ class SkvbcPersistenceTest(unittest.TestCase):
             log.log_message(message_type=f'Repeatedly restarting stale replica {stale} '
                   f'to with view change running in the background.')
 
-            await self._fetch_or_finish_state_transfer_while_crashing(
+            await skvbc._fetch_or_finish_state_transfer_while_crashing(
                 bft_network=bft_network,
                 up_to_date_node=random.choice(stable_replicas),
                 stale_node=stale,
